- rewritten `TooltipTest` in typescript - refactored/optimized `TooltipTest.tsx`
