+++ apps/src/componentLibrary/tooltip/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.3.0]()
+
+- rewritten `TooltipTest` in typescript
+- refactored/optimized `TooltipTest.tsx`
+
 ## [0.2.3](https://github.com/code-dot-org/code-dot-org/pull/59328)
 
 * minor types update
