+++ apps/src/componentLibrary/tooltip/_Tooltip.tsx
@@ -57,7 +57,7 @@ export const TooltipOverlay: React.FunctionComponent<TooltipOverlayProps> = ({
  *  * (✔) implementation of component approved by design team;
  *  * (✔) has storybook, covered with stories and documentation;
  *  * (✔) has tests: test every prop, every state and every interaction that's js related;
- *  * (see apps/test/unit/componentLibrary/TooltipTest.jsx)
+ *  * (see apps/test/unit/componentLibrary/TooltipTest.tsx)
  *  * (?) passes accessibility checks;
  *
  * ###  Status: ```Ready for dev```
